+++ library/network/packets/direction_ack.h
@@ -15,10 +15,10 @@ class DIRECTION_ACK_PACKET_HANDLER : public PACKET_HANDLER
 {
 public:
     struct DATA {
-        u32  x;
-        u32  y;
-        u32  z;
-        u32  UserId;
+        float  x;
+        float  y;
+        float  z;
+        u32    UserId;
     };
 
 public:
